Extract JSX related print
